+++ src/types/intro.interface.tsx
@@ -1,0 +1,4 @@
+export type IntroType = {
+  id: number;
+  intro: string;
+};
